rename render js files for consistency
